refactor: Se reorganiza hooks y service de users
